[test] move test to selector.swift
